+++ tests/v2/import_export_args.spec.js
@@ -81,17 +81,17 @@ test('Import/export arguments [v2]', async ({ page, workflow }) => {
 		await workflow.addUserTask(nonParallelTaskWithoutArgsSchema);
 		await workflow.selectTask(nonParallelTaskWithoutArgsSchema);
 		await page.getByRole('button', { name: 'Add property' }).click();
-		await page.getByPlaceholder('Arg name').click();
-		await page.getByPlaceholder('Arg name').fill('key_non_parallel');
-		await page.getByPlaceholder('Argument default value').fill('value_non_parallel');
-		await page.getByLabel('Save argument').click();
+		await page.getByPlaceholder('Argument name').click();
+		await page.getByPlaceholder('Argument name').fill('key_non_parallel');
+		await page.getByPlaceholder('Argument value').fill('value_non_parallel');
+		await page.getByRole('button', { name: 'Save changes' }).click();
 
 		const { file, data } = await exportArgs(page, nonParallelTaskWithoutArgsSchema);
 		expect(data.args_non_parallel.key_non_parallel).toEqual('value_non_parallel');
 		expect(data.args_parallel).toEqual(null);
 		const newData = { args_non_parallel: { key_non_parallel: 'value_non_parallel-updated' } };
 		await importValidArgs(page, file, newData);
-		await page.getByText('value_non_parallel-updated').waitFor();
+		await page.getByText('Arguments changes saved successfully').waitFor();
 		await workflow.removeCurrentTask();
 	});
 
@@ -99,17 +99,17 @@ test('Import/export arguments [v2]', async ({ page, workflow }) => {
 		await workflow.addUserTask(parallelTaskWithoutArgsSchema);
 		await workflow.selectTask(parallelTaskWithoutArgsSchema);
 		await page.getByRole('button', { name: 'Add property' }).click();
-		await page.getByPlaceholder('Arg name').click();
-		await page.getByPlaceholder('Arg name').fill('key_parallel');
-		await page.getByPlaceholder('Argument default value').fill('value_parallel');
-		await page.getByLabel('Save argument').click();
+		await page.getByPlaceholder('Argument name').click();
+		await page.getByPlaceholder('Argument name').fill('key_parallel');
+		await page.getByPlaceholder('Argument value').fill('value_parallel');
+		await page.getByRole('button', { name: 'Save changes' }).click();
 
 		const { file, data } = await exportArgs(page, parallelTaskWithoutArgsSchema);
 		expect(data.args_non_parallel).toEqual(null);
 		expect(data.args_parallel.key_parallel).toEqual('value_parallel');
 		const newData = { args_parallel: { key_parallel: 'value_parallel-updated' } };
 		await importValidArgs(page, file, newData);
-		await page.getByText('value_parallel-updated').waitFor();
+		await page.getByText('Arguments changes saved successfully').waitFor();
 		await workflow.removeCurrentTask();
 	});
 
@@ -117,14 +117,12 @@ test('Import/export arguments [v2]', async ({ page, workflow }) => {
 		await workflow.addUserTask(compoundTaskWithoutArgsSchema);
 		await workflow.selectTask(compoundTaskWithoutArgsSchema);
 		await page.getByRole('button', { name: 'Add property' }).first().click();
-		await page.getByPlaceholder('Arg name').fill('key_non_parallel');
-		await page.getByPlaceholder('Argument default value').fill('value_non_parallel');
-		await page.getByLabel('Save argument').click();
-		await page.getByText('Arguments changes saved successfully').waitFor();
+		await page.getByPlaceholder('Argument name').fill('key_non_parallel');
+		await page.getByPlaceholder('Argument value').fill('value_non_parallel');
 		await page.getByRole('button', { name: 'Add property' }).nth(1).click();
-		await page.getByPlaceholder('Arg name').fill('key_parallel');
-		await page.getByPlaceholder('Argument default value').fill('value_parallel');
-		await page.getByLabel('Save argument').click();
+		await page.getByPlaceholder('Argument name').nth(1).fill('key_parallel');
+		await page.getByPlaceholder('Argument value').nth(1).fill('value_parallel');
+		await page.getByRole('button', { name: 'Save changes' }).click();
 		await page.getByText('Arguments changes saved successfully').waitFor();
 
 		const { file, data } = await exportArgs(page, compoundTaskWithoutArgsSchema);
@@ -135,8 +133,7 @@ test('Import/export arguments [v2]', async ({ page, workflow }) => {
 			args_parallel: { key_parallel: 'value_parallel-updated' }
 		};
 		await importValidArgs(page, file, newData);
-		await page.getByText('value_non_parallel-updated').waitFor();
-		await page.getByText('value_parallel-updated').waitFor();
+		await page.getByText('Arguments changes saved successfully').waitFor();
 		await workflow.removeCurrentTask();
 	});